[FIX] Remove checking of column types from SQLite (#95)
SQLite does not have static types for columns and some columns may
detect incorrectly and fail this check. For example, a random salt is
generated as a hex string. This string can contain only decimal numbers.
In this case the column type returned will be an integer and not a
string.